Changes to exit session when user logs out

+++ src/app/modules/home/components/home.component.ts
@@ -109,8 +109,15 @@ export class HomeComponent implements OnInit, AfterViewInit {
     });
     dialogRef.afterClosed().subscribe(dialogResult => {
       if (dialogResult == true) {
-        sessionStorage.clear();
-        this.router.navigate(['/landing']);
+        this.userService.logoutFromBrokerage('FYERS').subscribe(result => {
+          if(result){
+            sessionStorage.clear();
+            this.router.navigate(['/landing']);
+          }
+        }, err => {
+          console.log("Error while logging out "+JSON.stringify(err));
+          this.toastr.error('Failed to logout', 'Error');
+        });
       }
     });
   }
@@ -130,7 +137,7 @@ export class HomeComponent implements OnInit, AfterViewInit {
           this.router.navigate(['/landing']);
           }
         }, err => {
-          console.log("Error while exiting session "+JSON.stringify(err))
+          console.log("Error while exiting session "+JSON.stringify(err));
           this.toastr.error('Failed to exit session', 'Error');
         });
       }
